feat: moves config logic into new workspace
- creates new `src/config` workspace
- updates root `README.md` licenses overview
- moves the `get`, `set`, `list`, `del` functions from
  `src/cli-sdk/src/commands/config` into new workspace
- moves core logic tests for commands into new workspace
- updates new workspace `README.md` for examples
