+++ src/cli-sdk/package.json
@@ -22,6 +22,7 @@
     }
   },
   "dependencies": {
+    "@vltpkg/config": "workspace:*",
     "@vltpkg/dep-id": "workspace:*",
     "@vltpkg/dot-prop": "workspace:*",
     "@vltpkg/error-cause": "workspace:*",
