+++ src/config/typedoc.mjs
@@ -1,0 +1,2 @@
+import config from '../../www/docs/typedoc.workspace.mjs'
+export default config(import.meta.dirname)
